+++ README.md
@@ -28,7 +28,7 @@
 
 - [x]  사용자 이름 문자 검증
 - [x]  사용자가 한명인 경우
-- [x]  구분자 사이에 공백이 들어올 수 있다
+- [x]  구분자 사이에 공백이 들어올 수 있다.
 - [x]  구분자가 맨처음 맨 끝에 들어올 수 없다.
-- [ ]  사다리 높이는 숫자
+- [x]  사다리 높이는 숫자만 들어올 수 있다.
 - [x]  입력받을 수 있는 사용자 이름은 최대 50명이다.
