+++ src/main/java/domain/Bridge.java
@@ -3,7 +3,6 @@ package domain;
 import java.util.Arrays;
 
 public enum Bridge {
-
     BRIDGE(true),
     NON_BRIDGE(false);
 
@@ -21,7 +20,6 @@ public enum Bridge {
         return Arrays.stream(values())
                 .filter(value -> value.bridgeType == bridgeType)
                 .findAny()
-                .orElseThrow();
+                .orElseThrow(()-> new IllegalArgumentException("bridgeType이 존재하지 않습니다."));
     }
-
 }
