+++ src/main/java/domain/Ladder.java
@@ -9,16 +9,18 @@ import java.util.stream.IntStream;
 public class Ladder {
     private final List<Line> ladder = new ArrayList<>();
 
-    public Ladder(final int maxHeight, final int personCount) {
-        validateMaxHeight(maxHeight);
+    public Ladder(final int height, final int personCount) {
+        validateMaxHeight(height);
 
-        IntStream.range(0, maxHeight)
+        IntStream.range(0, height)
                 .forEach(iterator -> ladder.add(new Line(personCount, new RandomGenerator())));
     }
 
-    private static void validateMaxHeight(int maxHeight) {
-        if (maxHeight > 100) {
-            throw new IllegalArgumentException(String.format("입력된 값: %d, 사다리 높이는 최대 100입니다.", maxHeight));
+    private void validateMaxHeight(int height) {
+        final int MAX_HEIGHT = 100;
+
+        if (height > MAX_HEIGHT) {
+            throw new IllegalArgumentException(String.format("입력된 값: %d, 사다리 높이는 최대 %d입니다.", height, MAX_HEIGHT));
         }
     }
