+++ src/main/java/domain/UserName.java
@@ -3,13 +3,13 @@ package domain;
 import java.util.regex.Pattern;
 
 public class UserName {
+    private final Pattern SPECIAL_CHARACTER = Pattern.compile("[ !@#$%^&*().?\":{}|<>]");
     private final String userName;
 
     public UserName(final String userName) {
         validate(userName);
         this.userName = userName;
     }
-
 
     private void validate(final String userName) {
         validateLength(userName);
@@ -30,9 +30,7 @@ public class UserName {
     }
 
     private void validateSpecialCharacter(final String userName) {
-        final Pattern specialCharacter = Pattern.compile("[ !@#$%^&*().?\":{}|<>]");
-
-        if (specialCharacter.matcher(userName).find()) {
+        if (SPECIAL_CHARACTER.matcher(userName).find()) {
             throw new IllegalArgumentException(String.format("입력 된 값: %s, 특수기호는 들어갈 수 없습니다.", userName));
         }
     }
