+++ src/main/java/domain/Users.java
@@ -3,7 +3,6 @@ package domain;
 import java.util.List;
 
 public class Users {
-
     private final List<UserName> users;
 
     public Users(final List<String> userNames) {
@@ -30,7 +29,7 @@ public class Users {
     }
 
     private void validate(final List<String> userNames) {
-        if (userNames.size() == 1) {
+        if (userNames.size() <= 1) {
             throw new IllegalArgumentException(String.format("입력된 값: %d, 사용자는 두명 이상이여야 합니다.", userNames.size()));
         }
         if (userNames.size() > 50) {
