+++ src/main/java/utils/Generator.java
@@ -1,6 +1,5 @@
 package utils;
 
 public interface Generator {
-
     boolean generate();
 }
