+++ src/main/java/utils/RandomGenerator.java
@@ -3,7 +3,6 @@ package utils;
 import java.util.Random;
 
 public class RandomGenerator implements Generator {
-
     private static final Random RANDOM = new Random();
 
     @Override
