+++ src/test/java/mock/TrueGenerator.java
@@ -1,4 +1,6 @@
-package utils;
+package mock;
+
+import utils.Generator;
 
 public class TrueGenerator implements Generator {
     @Override
